requirements.txt.txt
pandas 
sqlalchemy 
pymysql
python-dotenv

